Created Day 7 testers & classes, did last testers except in Index
